backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
google-generativeai>=0.3.1
PyPDF2>=3.0.0
python-docx>=1.0.0